Extract validate query runner partial
